FCA 11-07-2023

+++ ApiPortal/ApiPortal/ViewModelsPortal/DashboardComprasVm.cs
@@ -3,7 +3,7 @@
     public class DashboardComprasVm
     {
         public int CantidadDocumentos { get; set; }
-        public int MontoTotal { get; set; }
+        public decimal MontoTotal { get; set; }
         public string Tipo { get; set; }
     }
 }
